Start moving logic to seperate services
