+++ backend/src/services/kv_storage_service.rs
@@ -6,17 +6,10 @@ use super::{
 };
 use crate::{
     adapters::{errors::GenericStorageError, KVStorageAdapter},
-    error::{
-        ApplyRevisionError, ApproveRevisionError, CreateConfigError, CreateConfigInstanceError,
-        CreateLabelError, CreatePasswordResetLinkError, CreateProjectError, DeleteConfigError,
-        DeleteConfigInstanceError, ResetPasswordError, RollbackRevisionError,
-        SaveConfigInstanceError,
-    },
+    error::{CreateLabelError, CreatePasswordResetLinkError, ResetPasswordError},
     model::{
-        ConfigInstance, ConfigInstanceChange, ConfigInstanceRevision, LabelType,
-        RevisionReviewState, YakManApiKey, YakManConfig, YakManLabel, YakManPassword,
-        YakManPasswordResetLink, YakManProject, YakManPublicPasswordResetLink, YakManRole,
-        YakManUser, YakManUserDetails,
+        LabelType, YakManApiKey, YakManPassword, YakManPasswordResetLink,
+        YakManPublicPasswordResetLink, YakManRole, YakManUser, YakManUserDetails,
     },
 };
 use argon2::{
@@ -39,50 +32,6 @@ pub struct KVStorageService {
 
 #[async_trait]
 impl StorageService for KVStorageService {
-    async fn get_projects(&self) -> Result<Vec<YakManProject>, GenericStorageError> {
-        return Ok(self.adapter.get_projects().await?);
-    }
-
-    async fn get_config(
-        &self,
-        config_name: &str,
-    ) -> Result<Option<YakManConfig>, GenericStorageError> {
-        let c = self.adapter.get_configs().await?;
-        return Ok(c.into_iter().find(|c| c.name == config_name && !c.hidden));
-    }
-
-    async fn create_project(&self, project_name: &str) -> Result<String, CreateProjectError> {
-        let mut projects = self.adapter.get_projects().await?;
-
-        // Prevent duplicates
-        for prj in &projects {
-            if &prj.name == &project_name {
-                return Err(CreateProjectError::DuplicateNameError {
-                    name: String::from(project_name),
-                });
-            }
-        }
-
-        let project_uuid = Uuid::new_v4();
-
-        projects.push(YakManProject {
-            name: String::from(project_name),
-            uuid: project_uuid.to_string(),
-        });
-
-        self.adapter.save_projects(projects).await?;
-
-        return Ok(project_uuid.to_string());
-    }
-
-    async fn get_visible_configs(
-        &self,
-        project_uuid: Option<String>,
-    ) -> Result<Vec<YakManConfig>, GenericStorageError> {
-        let configs = self.get_all_configs(project_uuid).await?;
-        return Ok(configs.into_iter().filter(|c| !c.hidden).collect());
-    }
-
     async fn get_labels(&self) -> Result<Vec<LabelType>, GenericStorageError> {
         return Ok(self.adapter.get_labels().await?);
     }
@@ -116,171 +65,6 @@ impl StorageService for KVStorageService {
         return Ok(());
     }
 
-    async fn create_config_instance(
-        &self,
-        config_name: &str,
-        labels: Vec<YakManLabel>,
-        data: &str,
-        content_type: Option<String>,
-        creator_uuid: &str,
-    ) -> Result<String, CreateConfigInstanceError> {
-        if let Some(mut instances) = self.adapter.get_instance_metadata(config_name).await? {
-            let instance = generate_instance_id();
-            let revision_key: String = generate_revision_id();
-            let data_key = Uuid::new_v4().to_string();
-            let now = Utc::now().timestamp_millis();
-
-            if !self.validate_labels(&labels).await? {
-                return Err(CreateConfigInstanceError::InvalidLabel);
-            }
-
-            // Create new file with data
-            self.adapter
-                .save_instance_data(config_name, &data_key, data)
-                .await?;
-
-            // Create revision
-            let revision = ConfigInstanceRevision {
-                revision: String::from(&revision_key),
-                data_key: String::from(&data_key),
-                labels: labels,
-                timestamp_ms: now,
-                review_state: RevisionReviewState::Approved,
-                reviewed_by_uuid: Some(creator_uuid.to_string()),
-                review_timestamp_ms: Some(now),
-                submitted_by_uuid: creator_uuid.to_string(),
-                submit_timestamp_ms: now,
-                content_type: content_type.unwrap_or(String::from("text/plain")),
-            };
-            self.adapter.save_revision(config_name, &revision).await?;
-
-            // Add new instance to instances and update the instance metadata
-            instances.push(ConfigInstance {
-                config_name: config_name.to_string(),
-                instance: instance.to_string(),
-                labels: revision.labels,
-                current_revision: revision.revision.clone(),
-                pending_revision: None,
-                revisions: vec![revision.revision.clone()],
-                changelog: vec![ConfigInstanceChange {
-                    timestamp_ms: now,
-                    previous_revision: None,
-                    new_revision: revision.revision,
-                    applied_by_uuid: creator_uuid.to_string(),
-                }],
-            });
-            self.adapter
-                .save_instance_metadata(config_name, instances)
-                .await?;
-            info!("Update instance metadata for config: {}", config_name);
-
-            return Ok(instance);
-        }
-
-        return Err(CreateConfigInstanceError::NoConfigFound);
-    }
-
-    async fn create_config(
-        &self,
-        config_name: &str,
-        project_uuid: &str,
-    ) -> Result<(), CreateConfigError> {
-        let mut configs = self
-            .get_all_configs(None)
-            .await
-            .map_err(|_| CreateConfigError::storage_error("Failed to load configs"))?;
-
-        let mut config = configs.iter_mut().find(|config| config.name == config_name);
-
-        match &mut config {
-            Some(&mut ref mut config) => {
-                if !config.hidden {
-                    return Err(CreateConfigError::duplicate_config(config_name));
-                }
-
-                info!("Config '{config_name}' already exists, unhiding it");
-
-                // Config already exists, just unhide it
-                config.hidden = false;
-                self.adapter.save_configs(configs).await.map_err(|_| {
-                    CreateConfigError::storage_error("Failed to update configs file")
-                })?;
-                return Ok(());
-            }
-            None => (),
-        }
-
-        configs.push(YakManConfig {
-            name: String::from(config_name),
-            project_uuid: String::from(project_uuid),
-            hidden: false,
-        });
-
-        // Create instance metadata file
-        self.adapter
-            .save_instance_metadata(config_name, vec![])
-            .await
-            .map_err(|_| CreateConfigError::storage_error("Failed to save instance metadata"))?;
-
-        // Create config instances directory
-        self.adapter
-            .prepare_config_instance_storage(config_name)
-            .await
-            .map_err(|_| {
-                CreateConfigError::storage_error("Failed to create instances directory")
-            })?;
-
-        // Create config revisions directory
-        self.adapter
-            .prepare_revision_instance_storage(config_name)
-            .await
-            .map_err(|_| {
-                CreateConfigError::storage_error("Failed to create revisions directory")
-            })?;
-
-        // Add config to base config file
-        self.adapter
-            .save_configs(configs)
-            .await
-            .map_err(|_| CreateConfigError::storage_error("Failed to update configs file"))?;
-
-        Ok(())
-    }
-
-    async fn delete_config(&self, config_name: &str) -> Result<(), DeleteConfigError> {
-        let mut configs = self.get_visible_configs(None).await?;
-
-        if let Some(config) = configs
-            .iter_mut()
-            .find(|config| config.name == config_name && !config.hidden)
-        {
-            config.hidden = true;
-            self.adapter.save_configs(configs).await?;
-            return Ok(());
-        }
-
-        return Err(DeleteConfigError::ConfigDoesNotExistError);
-    }
-
-    async fn get_config_instance_metadata(
-        &self,
-        config_name: &str,
-    ) -> Result<Option<Vec<ConfigInstance>>, GenericStorageError> {
-        return Ok(self.adapter.get_instance_metadata(config_name).await?);
-    }
-
-    async fn get_config_instance(
-        &self,
-        config_name: &str,
-        instance: &str,
-    ) -> Result<Option<ConfigInstance>, GenericStorageError> {
-        let instances = self.get_config_instance_metadata(config_name).await?;
-        return match instances {
-            Some(instances) => Ok(instances.into_iter().find(|inst| inst.instance == instance)),
-            None => Ok(None),
-        };
-    }
-
     async fn get_config_data(
         &self,
         config_name: &str,
@@ -303,95 +87,6 @@ impl StorageService for KVStorageService {
         return Ok(None);
     }
 
-    async fn submit_new_instance_revision(
-        &self,
-        config_name: &str,
-        instance: &str,
-        labels: Vec<YakManLabel>,
-        data: &str,
-        content_type: Option<String>,
-        submitted_by_uuid: &str,
-    ) -> Result<String, SaveConfigInstanceError> {
-        let mut instances = self
-            .adapter
-            .get_instance_metadata(config_name)
-            .await?
-            .ok_or(SaveConfigInstanceError::InvalidConfig)?;
-
-        let instance = instances
-            .iter_mut()
-            .find(|inst| inst.instance == instance)
-            .ok_or(SaveConfigInstanceError::InvalidInstance)?;
-
-        if !self.validate_labels(&labels).await? {
-            return Err(SaveConfigInstanceError::InvalidLabel);
-        }
-
-        let revision_key = generate_revision_id();
-        let data_key = Uuid::new_v4().to_string();
-
-        // Create new file with data
-        self.adapter
-            .save_instance_data(config_name, &data_key, data)
-            .await?;
-
-        // Create revision
-        let now = Utc::now().timestamp_millis();
-        let revision = ConfigInstanceRevision {
-            revision: String::from(&revision_key),
-            data_key: String::from(&data_key),
-            labels: labels,
-            timestamp_ms: now,
-            review_state: RevisionReviewState::Pending,
-            reviewed_by_uuid: None,
-            review_timestamp_ms: None,
-            submitted_by_uuid: submitted_by_uuid.to_string(),
-            submit_timestamp_ms: now,
-            content_type: content_type.unwrap_or(String::from("text/plain")),
-        };
-        self.adapter.save_revision(config_name, &revision).await?;
-
-        // Update instance data
-        instance.pending_revision = Some(String::from(&revision.revision));
-        instance.revisions.push(String::from(&revision.revision));
-
-        self.adapter
-            .save_instance_metadata(config_name, instances)
-            .await?;
-
-        log::info!("Updated instance metadata for config: {config_name}");
-
-        return Ok(revision_key);
-    }
-
-    async fn get_instance_revisions(
-        &self,
-        config_name: &str,
-        instance: &str,
-    ) -> Result<Option<Vec<ConfigInstanceRevision>>, GenericStorageError> {
-        let instances = match self.get_config_instance_metadata(&config_name).await? {
-            Some(value) => value,
-            None => return Ok(None),
-        };
-
-        let instance = match instances.iter().find(|inst| inst.instance == instance) {
-            Some(value) => value,
-            None => return Ok(None),
-        };
-
-        info!("found {} revisions", instance.revisions.len());
-
-        let mut revisions: Vec<ConfigInstanceRevision> = vec![];
-
-        for rev in instance.revisions.iter() {
-            if let Some(revision) = self.adapter.get_revision(config_name, &rev).await? {
-                revisions.push(revision);
-            }
-        }
-
-        return Ok(Some(revisions));
-    }
-
     /// Returns a tuple of (data, content_type)
     async fn get_data_by_revision(
         &self,
@@ -407,210 +102,6 @@ impl StorageService for KVStorageService {
         }
         info!("Fetching revision not found");
         return Ok(None);
-    }
-
-    async fn approve_instance_revision(
-        &self,
-        config_name: &str,
-        instance: &str,
-        revision: &str,
-        approved_uuid: &str,
-    ) -> Result<(), ApproveRevisionError> {
-        let mut metadata = match self.get_config_instance_metadata(config_name).await? {
-            Some(metadata) => metadata,
-            None => return Err(ApproveRevisionError::InvalidConfig),
-        };
-
-        let instance = match metadata.iter_mut().find(|i| i.instance == instance) {
-            Some(instance) => instance,
-            None => return Err(ApproveRevisionError::InvalidInstance),
-        };
-
-        // Verify instance is the pending revision
-        if let Some(pending_revision) = &instance.pending_revision {
-            if pending_revision != revision {
-                return Err(ApproveRevisionError::InvalidRevision);
-            }
-        } else {
-            return Err(ApproveRevisionError::InvalidRevision);
-        }
-
-        let mut revision_data = match self.adapter.get_revision(config_name, revision).await.ok() {
-            Some(Some(revision_data)) => revision_data,
-            None | Some(None) => return Err(ApproveRevisionError::InvalidRevision),
-        };
-
-        let now = Utc::now().timestamp_millis();
-        revision_data.review_state = RevisionReviewState::Approved;
-        revision_data.reviewed_by_uuid = Some(approved_uuid.to_string());
-        revision_data.review_timestamp_ms = Some(now);
-        self.adapter
-            .save_revision(config_name, &revision_data)
-            .await?;
-
-        if !instance.revisions.contains(&String::from(revision)) {
-            instance.revisions.push(String::from(revision));
-        }
-
-        self.adapter
-            .save_instance_metadata(config_name, metadata)
-            .await?;
-
-        return Ok(());
-    }
-
-    async fn apply_instance_revision(
-        &self,
-        config_name: &str,
-        instance: &str,
-        revision: &str,
-        applied_by_uuid: &str,
-    ) -> Result<(), ApplyRevisionError> {
-        let mut metadata = match self.get_config_instance_metadata(config_name).await? {
-            Some(metadata) => metadata,
-            None => return Err(ApplyRevisionError::InvalidConfig),
-        };
-
-        let instance = match metadata.iter_mut().find(|i| i.instance == instance) {
-            Some(instance) => instance,
-            None => return Err(ApplyRevisionError::InvalidInstance),
-        };
-
-        // Verify instance is the pending revision
-        if let Some(pending_revision) = &instance.pending_revision {
-            if pending_revision != revision {
-                return Err(ApplyRevisionError::InvalidRevision);
-            }
-        } else {
-            return Err(ApplyRevisionError::InvalidRevision);
-        }
-
-        let revision_data = match self.adapter.get_revision(config_name, revision).await.ok() {
-            Some(Some(revision_data)) => revision_data,
-            None | Some(None) => return Err(ApplyRevisionError::InvalidRevision),
-        };
-
-        if revision_data.review_state != RevisionReviewState::Approved {
-            return Err(ApplyRevisionError::NotApproved);
-        }
-
-        let now = Utc::now().timestamp_millis();
-        instance.changelog.push(ConfigInstanceChange {
-            timestamp_ms: now,
-            previous_revision: Some(instance.current_revision.clone()),
-            new_revision: String::from(revision),
-            applied_by_uuid: String::from(applied_by_uuid),
-        });
-        instance.current_revision = String::from(revision);
-        instance.pending_revision = None;
-        instance.labels = revision_data.labels;
-
-        if !instance.revisions.contains(&String::from(revision)) {
-            instance.revisions.push(String::from(revision));
-        }
-
-        self.adapter
-            .save_instance_metadata(config_name, metadata)
-            .await?;
-
-        return Ok(());
-    }
-
-    async fn reject_instance_revision(
-        &self,
-        config_name: &str,
-        instance: &str,
-        revision: &str,
-        rejected_by_uuid: &str,
-    ) -> Result<(), ApplyRevisionError> {
-        let mut metadata = match self.get_config_instance_metadata(config_name).await? {
-            Some(metadata) => metadata,
-            None => return Err(ApplyRevisionError::InvalidConfig),
-        };
-
-        let instance = match metadata.iter_mut().find(|i| i.instance == instance) {
-            Some(instance) => instance,
-            None => return Err(ApplyRevisionError::InvalidInstance),
-        };
-
-        let mut revision_data = match self.adapter.get_revision(config_name, revision).await.ok() {
-            Some(Some(revision_data)) => revision_data,
-            None | Some(None) => return Err(ApplyRevisionError::InvalidRevision),
-        };
-
-        let now = Utc::now().timestamp_millis();
-        revision_data.review_state = RevisionReviewState::Rejected;
-        revision_data.reviewed_by_uuid = Some(rejected_by_uuid.to_string());
-        revision_data.review_timestamp_ms = Some(now);
-
-        instance.pending_revision = None;
-
-        if let Some(index) = instance.revisions.iter().position(|x| *x == revision) {
-            instance.revisions.remove(index);
-        }
-
-        self.adapter
-            .save_revision(config_name, &revision_data)
-            .await?;
-
-        self.adapter
-            .save_instance_metadata(config_name, metadata)
-            .await?;
-
-        return Ok(());
-    }
-
-    async fn rollback_instance_revision(
-        &self,
-        config_name: &str,
-        instance: &str,
-        revision: &str,
-        rollback_by_uuid: &str,
-    ) -> Result<String, RollbackRevisionError> {
-        let mut instances = self
-            .adapter
-            .get_instance_metadata(config_name)
-            .await?
-            .ok_or(RollbackRevisionError::InvalidConfig)?;
-
-        let instance = instances
-            .iter_mut()
-            .find(|inst| inst.instance == instance)
-            .ok_or(RollbackRevisionError::InvalidInstance)?;
-
-        let previous_revision = self
-            .adapter
-            .get_revision(&config_name, &revision)
-            .await?
-            .ok_or(RollbackRevisionError::InvalidRevision)?;
-
-        let revision_key = generate_revision_id();
-
-        // Create revision
-        let now = Utc::now().timestamp_millis();
-        let revision = ConfigInstanceRevision {
-            revision: String::from(&revision_key),
-            data_key: previous_revision.data_key,
-            labels: previous_revision.labels.clone(),
-            timestamp_ms: now,
-            review_state: RevisionReviewState::Pending,
-            reviewed_by_uuid: None,
-            review_timestamp_ms: Some(now),
-            submitted_by_uuid: rollback_by_uuid.to_string(),
-            submit_timestamp_ms: now,
-            content_type: previous_revision.content_type,
-        };
-        self.adapter.save_revision(config_name, &revision).await?;
-
-        // Update instance data
-        instance.pending_revision = Some(String::from(&revision.revision));
-        instance.revisions.push(String::from(&revision.revision));
-
-        self.adapter
-            .save_instance_metadata(config_name, instances)
-            .await?;
-        log::info!("Updated instance metadata for config: {config_name}");
-        return Ok(revision_key);
     }
 
     async fn initialize_storage(&self) -> Result<(), GenericStorageError> {
@@ -772,41 +263,6 @@ impl StorageService for KVStorageService {
         return self.adapter.save_api_keys(api_keys).await;
     }
 
-    async fn delete_instance(
-        &self,
-        config_name: &str,
-        instance: &str,
-    ) -> Result<(), DeleteConfigInstanceError> {
-        let instances = self
-            .adapter
-            .get_instance_metadata(config_name)
-            .await?
-            .ok_or(DeleteConfigInstanceError::InvalidConfig)?;
-
-        let config_instance = instances
-            .iter()
-            .find(|i| i.instance == instance)
-            .ok_or(DeleteConfigInstanceError::InvalidInstance)?
-            .clone();
-
-        let remaining_instances = instances
-            .into_iter()
-            .filter(|i| i.instance != instance)
-            .collect();
-
-        self.adapter
-            .save_instance_metadata(config_name, remaining_instances)
-            .await?;
-
-        for revision in config_instance.revisions {
-            if let Err(e) = self.adapter.delete_revision(config_name, &revision).await {
-                log::error!("Failed to delete revision ({revision}) {e:?}");
-            }
-        }
-
-        return Ok(());
-    }
-
     async fn get_password_by_email(
         &self,
         email: &str,
@@ -952,48 +408,6 @@ impl KVStorageService {
             api_key_hash_cache,
         }
     }
-
-    /// Gets all configs including hidden configs
-    async fn get_all_configs(
-        &self,
-        project_uuid: Option<String>,
-    ) -> Result<Vec<YakManConfig>, GenericStorageError> {
-        let configs = match project_uuid {
-            Some(project_uuid) => {
-                self.adapter
-                    .get_configs_by_project_uuid(project_uuid)
-                    .await?
-            }
-            None => self.adapter.get_configs().await?,
-        };
-        return Ok(configs);
-    }
-
-    /// Returns true if all labels exist and have valid values
-    async fn validate_labels(
-        &self,
-        labels: &Vec<YakManLabel>,
-    ) -> Result<bool, GenericStorageError> {
-        let all_labels = self.get_labels().await?;
-        for label in labels {
-            if let Some(label_type) = all_labels.iter().find(|l| l.name == label.label_type) {
-                if !label_type.options.iter().any(|opt| opt == &label.value) {
-                    return Ok(false);
-                }
-            } else {
-                return Ok(false);
-            }
-        }
-        return Ok(true);
-    }
-}
-
-fn generate_instance_id() -> String {
-    return format!("i{}", short_sha(&Uuid::new_v4().to_string()));
-}
-
-fn generate_revision_id() -> String {
-    return format!("r{}", short_sha(&Uuid::new_v4().to_string()));
 }
 
 /// Returns a 12 character string representation of a SHA256
@@ -1020,23 +434,5 @@ mod tests {
 
         let result = short_sha("ade10004-41df-4bf6-88b9-d768afab674f");
         assert_eq!(result, "8146205a8d27");
-    }
-
-    #[test]
-    fn test_generate_instance_id() {
-        for _i in 0..10 {
-            let result = generate_instance_id();
-            assert_eq!(13, result.len());
-            assert!(result.starts_with('i'));
-        }
-    }
-
-    #[test]
-    fn test_generate_revision_id() {
-        for _i in 0..10 {
-            let result = generate_revision_id();
-            assert_eq!(13, result.len());
-            assert!(result.starts_with('r'));
-        }
     }
 }
